requirements.txt
Flask==1.1.4
Flask-Injector==0.12.3
Flask-Script==2.0.6
flask-restplus==0.13.0
flask_accepts==0.17.4
waitress==1.4.4
PyYAML==5.3.1
config-client==0.11.0
Werkzeug==0.16.1
openpyxl==3.0.9
healthcheck==1.3.3
numpy==1.20.2
pandas==1.1.5
google-api-python-client==1.12.2
timeout-decorator==0.4.1
markupsafe==2.0.1